+++ CHANGELOG.md
@@ -7,7 +7,9 @@
 ### Added
 
 * Actions `chaosk8s.secret.actions.create_secret` and `chaosk8s.secret.actions.delete_secret`
+* Actions `chaosk8s.namespace.actions.create_namespace` and `chaosk8s.namespace.actions.delete_namespace`
 * Probes `chaosk8s.secret.probes.secret_exists`
+* Probes `chaosk8s.namespace.probes.namespace_exists`
 * Support the `NO_PROXY` environment variable
 
 ### Fixed